feat: Default TI ML features constant added.
The features in the ML data, are the signals of the Technical
Indicators. Which features will be used are being controlled by the
user. The Default set of features consists of all the 62 technical
indicators. A constant defined for this purpose in the tti.utils
package, constants module. Related to teh issue #23.